requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
absl-py==1.2.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.0.3
antlr4-python3-runtime==4.9.3
anyio==3.6.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backcall==0.2.0
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.0.4
cycler==0.11.0
cymem==2.0.6
Cython==0.29.32
dataclasses==0.6
decorator==5.1.1
docker-pycreds==0.4.0
effdet==0.2.4
entrypoints==0.4
fastai==2.5.6
fastapi==0.79.0
fastcore==1.3.29
fastdownload==0.0.7
fastprogress==1.0.3
fire==0.4.0
fonttools==4.34.4
frozenlist==1.3.1
fsspec==2022.7.1
fvcore==0.1.5.post20220512
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.10.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h11==0.13.0
icedata==0.5.1
icevision==0.12.0
idna==3.3
imageio==2.21.1
importlib-metadata==4.12.0
iopath==0.1.10
ipykernel==5.5.6
ipython==7.34.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver==1.4.4
langcodes==3.3.0
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
mmcv-full==1.3.17 -f https://download.openmmlab.com/mmcv/dist/cpu/torch1.10.0/index.html
mmdet==2.17.0
mmsegmentation==0.20.2
multidict==6.0.2
murmurhash==1.0.7
nest-asyncio==1.5.5
networkx==2.6.3
nose==1.3.7
numpy==1.21.6
oauthlib==3.2.0
omegaconf==2.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.1.2.30
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
portalocker==2.5.1
preshed==3.0.6
prettytable==3.3.0
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.4
pycocotools==2.0.4
pydantic==1.9.2
pyDeprecate==0.3.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.7.1
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
requests-oauthlib==1.3.1
resnest==0.0.6b20220813
rsa==4.9
sahi==0.10.4
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sentry-sdk==1.9.4
setproctitle==1.3.2
Shapely==1.8.2
shortuuid==1.0.9
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
spacy==3.4.1
spacy-legacy==3.0.9
spacy-loggers==1.0.3
srsly==2.4.4
starlette==0.19.1
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminaltables==3.1.10
thinc==8.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.7
torch==1.10.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
torchmetrics==0.9.3
torchtext==0.11.0 -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.11.1+cpu -f https://download.pytorch.org/whl/torch_stable.html
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typer==0.4.2
typing_extensions==4.1.1
urllib3==1.26.11
uvicorn==0.18.2
wandb==0.13.1
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.2
yacs==0.1.8
yapf==0.32.0
yarl==1.8.1
yolov5-icevision==6.0.0
zipp==3.8.1
gdown